reboot the RealBoy project
I wrote the original RealBoy emulator back in 2012-2013.

I've long wanted to rewrite it from scratch, reflecting more than
a decade of knowledge and experience. Finally I could spend an entire
month redesigning the whole thing with the same pedagogical aims in mind
as the original version.

Hopefully in can inspire people to develop love for the craft in the age
of AI.

Signed-off-by: Sergio Gómez <sergio.g.delreal@gmail.com>